refactor: /my/modify 페이지 전환
